Fix label `for` attribute missing form namespace
Previously, when a form includes a `namespace` and a label inside that
form has an explicit `for`, the rendered `for` attribute would not be
prefixed with the `namespace`.

The problem is that the label tag uses `add_default_name_and_id`, which
operates on the `id` option of the hash passed in. However, since the
label tag needs the `for` value set, `render` does a lot of extra work
to move the value from `id` to `for`, _sometimes_.

In this case, we could fix the condition for copying `id` to `for`,
however I think a better solution is to change `add_default_name_and_id`
to `add_default_name_and_field` and allow the field being set to be
configured. This simplifies the label's `render` by removing the need to
duplicate options and conditionally copy `id` into `for` at all.

Co-authored-by: abeidahmed <abeidahmed92@gmail.com>
Co-authored-by: Hartley McGuire <skipkayhil@gmail.com>